requirements.txt
flask
flask_sqlalchemy
Flask_session
pillow